feat: implement user-specific todo scoping and add logout functionality

+++ src/modules/todo/todo.services.js
@@ -1,8 +1,8 @@
 import { pool } from "../../config/db.js"
 
-export const getTodos = async () => {
+export const getTodos = async (userId) => {
     try {
-        const result = await pool.query("SELECT * FROM todos ORDER BY id DESC");
+        const result = await pool.query("SELECT * FROM todos WHERE user_id = $1 ORDER BY id DESC", [userId]);
         return result.rows;
     } catch (err) {
         console.error(err);
@@ -10,9 +10,9 @@ export const getTodos = async () => {
     }
 }
 
-export const createTodos = async (title) => {
+export const createTodos = async (title, userId) => {
     try {
-        const result = await pool.query("INSERT INTO todos (title) VALUES ($1) RETURNING *", [title]);
+        const result = await pool.query("INSERT INTO todos (title, user_id) VALUES ($1, $2) RETURNING *", [title, userId]);
         return result.rows[0];
     } catch (err) {
         console.error(err);
@@ -20,9 +20,9 @@ export const createTodos = async (title) => {
     }
 }
 
-export const deleteTodos = async (id) => {
+export const deleteTodos = async (id, userId) => {
     try {
-        const result = await pool.query("DELETE FROM todos where id=$1 RETURNING *", [id])
+        const result = await pool.query("DELETE FROM todos where id=$1 AND user_id = $2 RETURNING *", [id, userId])
         return result.rows[0];
     } catch (err) {
         console.error(err);
@@ -30,9 +30,9 @@ export const deleteTodos = async (id) => {
     }
 }
 
-export const updateTodosStatus = async (id, completed) => {
+export const updateTodosStatus = async (id, completed, userId) => {
     try {
-        const result = await pool.query("UPDATE todos SET completed = $1 WHERE id = $2", [completed, id])
+        const result = await pool.query("UPDATE todos SET completed = $1 WHERE id = $2 AND user_id = $3 RETURNING *", [completed, id, userId])
         return result.rows[0];
     } catch (err) {
         console.error(err);
@@ -40,9 +40,9 @@ export const updateTodosStatus = async (id, completed) => {
     }
 }
 
-export const updateTodos = async (id, title) => {
+export const updateTodos = async (id, title, userId) => {
     try {
-        const result = await pool.query("UPDATE todos SET title = $1 WHERE id = $2 RETURNING *", [title, id]);
+        const result = await pool.query("UPDATE todos SET title = $1 WHERE id = $2 AND user_id = $3 RETURNING *", [title, id, userId]);
         return result.rows[0];
     } catch (err) {
         console.error(err);
